refactor(10-03): split custom_auth/views.py into 5 domain modules (QA-03)
Moves all 9 view functions verbatim from the 911-line monolith into a
custom_auth/views/ package with domain-specific modules (D-04):

- oauth.py: google_exchange (AllowAny)
- registration.py: register_with_email (AllowAny)
- jwt_auth.py: get_jwt_token, refresh_jwt_token (AllowAny)
- profile.py: user_profile (IsAuthenticated), logout (AllowAny)
- password_reset.py: password_reset_request, password_reset_validate_token,
  password_reset_confirm (AllowAny)

__init__.py re-exports all 9 names so custom_auth/urls.py is unchanged (D-05).
Monolith deleted only after Docker pytest confirmed all 63 auth tests pass (D-06).

Also updates test mock patch paths from custom_auth.views.requests to
custom_auth.views.oauth.requests to match the new module location.

Full pytest suite: 195 passed (baseline unchanged).
